+++ packages/image-mapping/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sugarch/bc-image-mapping
 
+## 1.0.7
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @sugarch/bc-mod-hook-manager@0.2.12
+    -   @sugarch/bc-mod-utility@0.2.1
+
 ## 1.0.6
 
 ### Patch Changes