+++ packages/mod-hook-manager/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sugarch/bc-mod-hook-manager
 
+## 0.2.12
+
+### Patch Changes
+
+-   Add package info in dist files
+
 ## 0.2.11
 
 ### Patch Changes